fix(frontend.html): session expired in server, sync with client
Each user session is going to expire after 24hrs, But since I abused the
localStorage in Javascript, it never expires in client side.

So, if a request from a logged in page went failed, alert user and
redirect him/her to login page.

backend: add a little printfuck to help debug with db operation